+++ tests/test_numerapi.py
@@ -9,17 +9,19 @@ from numerapi.numerapi import NumerAPI
 import numerapi
 
 
-def test_get_competitions():
-    api = NumerAPI()
+@pytest.fixture(scope='function', name="api")
+def api_fixture():
+    api = NumerAPI(verbosity='DEBUG')
+    return api
 
-    # get all competions
+
+def test_get_competitions(api):
     res = api.get_competitions()
     assert isinstance(res, list)
     assert len(res) > 80
 
 
-def test_download_current_dataset():
-    api = NumerAPI()
+def test_download_current_dataset(api):
     path = api.download_current_dataset(unzip=True)
     assert os.path.exists(path)
 
@@ -28,28 +30,24 @@ def test_download_current_dataset():
     assert os.path.exists(os.path.join(directory, filename))
 
 
-def test_get_current_round():
-    api = NumerAPI()
+def test_get_current_round(api):
     current_round = api.get_current_round()
     assert current_round >= 82
 
 
-def test_raw_query():
-    api = NumerAPI()
+def test_raw_query(api):
     query = "query {dataset}"
     result = api.raw_query(query)
     assert isinstance(result, dict)
     assert "data" in result
 
 
-def test_get_leaderboard():
-    api = NumerAPI()
+def test_get_leaderboard(api):
     lb = api.get_leaderboard(67)
     assert len(lb) == 1425
 
 
-def test_stake():
-    api = NumerAPI()
+def test_stake(api):
     with pytest.raises(ValueError) as err:
         # while this won't work because we are not authorized, it still tells
         # us if the request is formatted correctly
@@ -58,22 +56,19 @@ def test_stake():
     assert "You must be authenticated" in str(err.value)
 
 
-def test_get_staking_leaderboard():
-    api = NumerAPI()
+def test_get_staking_leaderboard(api):
     stakes = api.get_staking_leaderboard(82)
     # 115 people staked that round
     assert len(stakes) == 115
 
 
-def test_get_submission_ids():
-    api = NumerAPI()
+def test_get_submission_ids(api):
     ids = api.get_submission_ids()
     assert len(ids) > 0
     assert isinstance(ids, dict)
 
 
-def test_error_handling():
-    api = NumerAPI()
+def test_error_handling(api):
     # String instead of Int
     with pytest.raises(ValueError):
         api.get_leaderboard("foo")
@@ -88,8 +83,7 @@ def test_error_handling():
         api.submission_status()
 
 
-def test_check_new_round():
-    api = NumerAPI()
+def test_check_new_round(api):
     with requests_mock.mock() as m:
         open_time = datetime.datetime.utcnow().replace(tzinfo=pytz.utc)
         data = {"data": {"rounds": [{"openTime": open_time.isoformat()}]}}