+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Notable changes to this project.
 - return proper Python data types, for example the NMR amounts are now
   floats and no longer strings
 - show progress bar while downloading dataset
+- general code cleanup & more tests
 
 ## [0.7.1] - 2017-12-29
 - fix import issues (py2 vs py3)